commit com algumas alterações para subir o PR
